+++ path/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add a "parse" filter to get path components
+- Add both "present" end "absent" tests to check path dict state
 
 ## [2.2.0] - 2026-01-02